Fix cast in assignment 13
The devision of highest_total by 4 will result in an integer.
Thus a cast to float will not yield the expected result.
Diving by a float value forces a float division.

+++ Assignment-13/assignment13.c
@@ -61,9 +61,8 @@ int main(void)
 
 	//initialize a new float variable which will be our average
 	//use the (float) operator to change the integer variable being divided into a float operation
-	float ppg = (float)(highest_total / 4);
+	float ppg = highest_total / 4.0f;
 	printf("Player #%d had the highest scoring average at %.2f points per game.\n", player_id, ppg);
 	
 	return 0;
-
 }
